Skyline: move UNIFI tests to TestConnected.dll, where they won't interfere with TC builds if there are server issues
