+++ pwiz_tools/Skyline/TestConnected/Results/RemoteApi/UnifiFunctionalTest.cs
@@ -1,4 +1,23 @@
-﻿using System.Linq;
+﻿/*
+ * Original author: Nicholas Shulman <nicksh .at. u.washington.edu>,
+ *                  MacCoss Lab, Department of Genome Sciences, UW
+ *
+ * Copyright 2018 University of Washington - Seattle, WA
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using pwiz.Skyline.Controls.Graphs;
 using pwiz.Skyline.FileUI;
@@ -7,7 +26,7 @@ using pwiz.Skyline.Model.Results.RemoteApi;
 using pwiz.Skyline.ToolsUI;
 using pwiz.SkylineTestUtil;
 
-namespace pwiz.SkylineTestFunctional
+namespace pwiz.SkylineTestConnected.Results.RemoteApi
 {
     [TestClass]
     public class UnifiFunctionalTest : AbstractFunctionalTestEx
