+++ pwiz_tools/Skyline/TestConnected/TestConnected.csproj
@@ -115,6 +115,11 @@
       <Private>True</Private>
     </Reference>
     <Reference Include="System.Design" />
+    <Reference Include="IdentityModel, Version=2.6.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\Shared\Lib\IdentityModel.dll</HintPath>
+    </Reference>
+    <Reference Include="System.Net.Http" />
     <Reference Include="System.Xml.Linq">
       <RequiredTargetFramework>3.5</RequiredTargetFramework>
     </Reference>
@@ -133,7 +138,9 @@
       <RequiredTargetFramework>3.0</RequiredTargetFramework>
     </Reference>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <None Include="Results\RemoteApi\UnifiFunctionalTest.zip" />
+  </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\Shared\BiblioSpec\BiblioSpec.csproj">
       <Project>{32CC9B1A-9442-4B56-AF34-FB47595278A1}</Project>
@@ -197,6 +204,8 @@
   <ItemGroup>
     <Compile Include="Results\RemoteApi\ChorusContentsTest.cs" />
     <Compile Include="Results\RemoteApi\ChromTaskListTest.cs" />
+    <Compile Include="Results\RemoteApi\UnifiFunctionalTest.cs" />
+    <Compile Include="Results\RemoteApi\UnifiTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Results\RemoteApi\DdaSmall.ChorusRequest.xml" />
